requirements.txt
accelerate>=0.16.0
datasets
torch
torchvision
tqdm
Pillow
diffusers
python-dotenv